fix(raijin): isolate obsolete runtime cleanup
