+++ yarn/plugin-essentials/sources/commands/set-version.command.ts
@@ -1,14 +1,14 @@
-import { BaseCommand }                     from '@yarnpkg/cli'
-import { Configuration }                   from '@yarnpkg/core'
-import { Command }                         from 'clipanion'
+import { BaseCommand }                       from '@yarnpkg/cli'
+import { Configuration }                     from '@yarnpkg/core'
+import { Command }                           from 'clipanion'
 
-import { assertInstalledRaijinRuntime }    from './set-version.runtime.js'
-import { cleanupLegacyRaijinRuntimeFiles } from './set-version.runtime.js'
-import { fetchRaijinRuntimeManifest }      from './set-version.runtime.js'
-import { installRaijinRuntime }            from './set-version.runtime.js'
-import { findPackageCwd }                  from './set-version.utils.js'
-import { portableToNativePath }            from './set-version.utils.js'
-import { preparePackageProjectBoundary }   from './set-version.utils.js'
+import { cleanupObsoleteRaijinRuntimeFiles } from './set-version.migration.js'
+import { assertInstalledRaijinRuntime }      from './set-version.runtime.js'
+import { fetchRaijinRuntimeManifest }        from './set-version.runtime.js'
+import { installRaijinRuntime }              from './set-version.runtime.js'
+import { findPackageCwd }                    from './set-version.utils.js'
+import { portableToNativePath }              from './set-version.utils.js'
+import { preparePackageProjectBoundary }     from './set-version.utils.js'
 
 export class SetVersionCommand extends BaseCommand {
   static paths = [['set', 'version', 'atls']]
@@ -42,7 +42,7 @@ export class SetVersionCommand extends BaseCommand {
       )
 
       await assertInstalledRaijinRuntime(updatedConfiguration, cwd, runtimeManifest)
-      await cleanupLegacyRaijinRuntimeFiles(cwd)
+      await cleanupObsoleteRaijinRuntimeFiles(cwd)
 
       const bumpArgs = ['up', '@atls/code-runtime']
       const bumpExitCode = await this.cli.run(bumpArgs, { cwd: cwd as typeof this.context.cwd })
